Project code review, add some architecture fixes and global code cleaning
