+++ calculator-credite/src/HOC/with-creditrate-service.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import { CreditRateServiceConsumer } from '../components/creditrate-service-context';
+
+const widthCreditRateService = (Wrapped) => {
+  return (props) => {
+    return (
+      <CreditRateServiceConsumer>
+        { 
+          (creditRateService) => 
+            <Wrapped {...props} creditRateService={creditRateService} />
+        }
+      </CreditRateServiceConsumer>
+    );
+  }
+};
+
+export default widthCreditRateService;
