--- calculator-credite/src/components/app-header/index.js
@@ -1,3 +1,0 @@
-import AppHeader from './app-header'
-
-export default AppHeader;
