--- calculator-credite/src/components/creditrate-service-context/creditrate-service-context.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const {
-  Provider: CreditRateServiceProvider,
-  Consumer: CreditRateServiceConsumer
-} = React.createContext();
-
-export {
-  CreditRateServiceProvider,
-  CreditRateServiceConsumer
-}
