--- calculator-credite/src/components/hoc/index.js
@@ -1,5 +1,0 @@
-import widthCreditRateService from './with-creditrate-service'
-
-export {
-  widthCreditRateService
-};
