--- calculator-credite/src/components/hoc/with-creditrate-service.js
@@ -1,20 +1,0 @@
-import React from 'react';
-import { CreditRateServiceConsumer} from '../creditrate-service-context';
-
-const widthCreditRateService = () =>  (Wrapped) => {
-  return(props)=>{
-    return(
-      <CreditRateServiceConsumer>
-        {
-          (creditRateService) => {
-            return (
-              <Wrapped {...props} creditRateService={creditRateService} />
-            )
-          }
-        }
-      </CreditRateServiceConsumer>
-    );
-  }
-};
-
-export default widthCreditRateService;
