--- calculator-credite/src/components/pages/app-calculator/app-calculator.css
@@ -1,129 +1,0 @@
-.table th {
-    vertical-align: middle;
-}
-
-.table span {
-    cursor: pointer;
-}
-
-.table span:hover {
-    color: blue;
-}
-
-.small-text {
-    color: #939393;
-    font-size: 12px;
-}
-
-.form-control-result {
-    font-weight: 600;
-    font-size: 18px;
-    margin: 0;
-}
-
-.resultCredit {
-    border: 2px solid #007bff;
-    border-radius: 5px;
-    padding: 15px 24px;
-}
-
-.resultCredit p {
-    font-weight: 600;
-}
-
-.resultCredit .big-text {
-    font-size: 30px;
-    font-weight:bold;
-}
-
-.center{
-    text-align: center;
-}
-
-input[type=range] {
-    -webkit-appearance: none; /* Hides the slider so that custom slider can be made */
-    width: 100%; /* Specific width is required for Firefox. */
-    background: transparent; /* Otherwise white in Chrome */
-    margin-top: 10px;
-}
-
-input[type=range]::-webkit-slider-thumb {
-    -webkit-appearance: none;
-}
-
-input[type=range]:focus {
-    outline: none; /* Removes the blue border. You should probably do some kind of focus styling for accessibility reasons though. */
-}
-
-input[type=range]::-ms-track {
-    width: 100%;
-    cursor: pointer;
-
-    /* Hides the slider so custom styles can be added */
-    background: transparent;
-    border-color: transparent;
-    color: transparent;
-}
-
-/* Special styling for WebKit/Blink */
-input[type=range]::-webkit-slider-thumb {
-    -webkit-appearance: none;
-    height: 20px;
-    width: 20px;
-    border-radius: 50%;
-    background: #ffe011;
-    cursor: pointer;
-    margin-top: -10px; /* You need to specify a margin in Chrome, but in Firefox and IE it is automatic */
-
-}
-
-/* All the same stuff for Firefox */
-input[type=range]::-moz-range-thumb {
-    height: 20px;
-    width: 20px;
-    border-radius: 50%;
-    background: #ffffff;
-    cursor: pointer;
-}
-
-/* All the same stuff for IE */
-input[type=range]::-ms-thumb {
-    box-shadow: 1px 1px 1px #000000, 0px 0px 1px #0d0d0d;
-    height: 20px;
-    width: 20px;
-    border-radius: 50%;
-    background: #ffffff;
-    cursor: pointer;
-}
-
-input[type=range]::-webkit-slider-runnable-track {
-    width: 100%;
-    height: 2px;
-    cursor: pointer;
-    background: #007bff;
-    border-radius: 1.3px;
-    border: 0.2px solid #010101;
-}
-
-input[type=range]:focus::-webkit-slider-runnable-track {
-    background: #007bff;
-}
-
-input[type=range]::-moz-range-track {
-    width: 100%;
-    height: 2px;
-    cursor: pointer;
-    background: #007bff;
-    border-radius: 1.3px;
-    border: 0.2px solid #010101;
-}
-
-input[type=range]::-ms-track {
-    width: 100%;
-    height: 2px;
-    cursor: pointer;
-    background: transparent;
-    border-color: transparent;
-    border-width: 16px 0;
-    color: transparent;
-}
