--- calculator-credite/src/components/pages/app-calculator/index.js
@@ -1,3 +1,0 @@
-import AppCalculator from './app-calculator';
-
-export default AppCalculator;
