--- calculator-credite/src/components/pages/app-home/index.js
@@ -1,3 +1,0 @@
-import AppHome from './app-home.js';
-
-export default AppHome;
