--- calculator-credite/src/components/pages/app-result/index.js
@@ -1,3 +1,0 @@
-import AppResult from './app-result'
-
-export default AppResult;
