--- calculator-credite/src/components/preloader/index.js
@@ -1,3 +1,0 @@
-import Preloader from './preloader'
-
-export default Preloader;
