+++ calculator-credite/src/containers/App/AppStyles.css
@@ -1,0 +1,8 @@
+.App {
+  margin: 50px auto;
+  max-width: 800px;
+}
+
+.row{
+  justify-content: center;
+}
