+++ calculator-credite/src/index.js
@@ -1,28 +1,25 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './components/app';
-/*import * as serviceWorker from './serviceWorker';*/
+import { render } from 'react-dom';
 import {BrowserRouter as Router} from 'react-router-dom';
-
 import {Provider} from 'react-redux';
+
+import App from './containers/App/App';
+
 import CreditRateService from './services/creditrate-service';
-import {CreditRateServiceProvider} from './components/creditrate-service-context';
+import { CreditRateServiceProvider } from './components/creditrate-service-context';
 import store from './store';
+
+import './index.css';
 
 const creditRateService = new CreditRateService();
 
-ReactDOM.render(
+render(
   <Provider store={store}>
     <CreditRateServiceProvider value={creditRateService}>
       <Router>
         <App />
       </Router>
     </CreditRateServiceProvider>
-  </Provider>
-  , document.getElementById('root'));
-
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: http://bit.ly/CRA-PWA
-/*serviceWorker.unregister();*/
+  </Provider>, 
+  document.getElementById('root')
+);
