+++ calculator-credite/src/pages/Result/Result.js
@@ -1,19 +1,19 @@
 import React from 'react';
-import {NavLink} from 'react-router-dom';
-import './app-result.css';
+import { NavLink } from 'react-router-dom';
 
-const AppResult = (props) => {
-// console.log(this.state);
+const Result = (props) => {
   if (props.location.query) {
-    const {result} = props.location.query;
-    console.log(result);
+    const { result } = props.location.query;
+
     return (
       <div>
         <div className="row">
           <h1>Результат</h1>
         </div>
+
         <div className="row">
           <h3>И так, на что Вы подписались</h3>
+
           <table className="table">
             <thead>
             <tr>
@@ -23,9 +23,9 @@ const AppResult = (props) => {
             </thead>
             <tbody>
             {
-              result.map((item, ind) => {
+              result.map((item, key) => {
                 return (
-                  <tr key={ind}>
+                  <tr key={key}>
                     <td>{item.name}</td>
                     <td>{item.value} {item.symbol}</td>
                   </tr>
@@ -34,17 +34,14 @@ const AppResult = (props) => {
             }
             </tbody>
           </table>
+
           <NavLink to="/calculator" className="badge badge-light mt-3">← Верунуться к выбору ставок</NavLink>
         </div>
       </div>
     );
   } else {
-    return (<div>
-      {props.history.push('/calculator')}
-    </div>);
+    return <div>{ props.history.push('/calculator') }</div>;
   }
-
-
 }
 
-export default AppResult;
+export default Result;
