+++ calculator-credite/src/reducers/index.js
@@ -1,21 +1,18 @@
-
 const initState = {
   creditrate: [],
   loading: true
 };
 
 const reducer = (state = initState, action) => {
-
-  switch(action.type){
+  switch (action.type) {
     case 'CREDITRATE_LOADED':
       return {
-        creditrate: action.payload,
+        creditrate: action.rate,
         loading: false
       };
     default:
       return state;
   }
-
 };
 
 export default reducer;
